fix the PostStore test suite after FluxC renaming
