+++ example/src/androidTest/java/org/wordpress/android/fluxc/release/ReleaseStack_SiteTest.java
@@ -70,7 +70,7 @@ public class ReleaseStack_SiteTest extends ReleaseStack_Base {
         RefreshSitesXMLRPCPayload payload = new RefreshSitesXMLRPCPayload();
         payload.username = BuildConfig.TEST_WPORG_USERNAME_SH_SIMPLE_CONTRIB;
         payload.password = BuildConfig.TEST_WPORG_PASSWORD_SH_SIMPLE_CONTRIB;
-        payload.url = BuildConfig.TEST_WPORG_URL_SH_SIMPLE_CONTRIB;
+        payload.url = BuildConfig.TEST_WPORG_URL_SH_SIMPLE_CONTRIB_ENDPOINT;
         mNextEvent = TEST_EVENTS.SITE_CHANGED;
         mCountDownLatch = new CountDownLatch(1);
         mDispatcher.dispatch(SiteActionBuilder.newFetchSitesXmlRpcAction(payload));
@@ -82,7 +82,7 @@ public class ReleaseStack_SiteTest extends ReleaseStack_Base {
         RefreshSitesXMLRPCPayload payload = new RefreshSitesXMLRPCPayload();
         payload.username = BuildConfig.TEST_WPORG_USERNAME_SH_MULTISITE;
         payload.password = BuildConfig.TEST_WPORG_PASSWORD_SH_MULTISITE;
-        payload.url = BuildConfig.TEST_WPORG_URL_SH_MULTISITE;
+        payload.url = BuildConfig.TEST_WPORG_URL_SH_MULTISITE_ENDPOINT;
         mNextEvent = TEST_EVENTS.SITE_CHANGED;
         mCountDownLatch = new CountDownLatch(1);
         mDispatcher.dispatch(SiteActionBuilder.newFetchSitesXmlRpcAction(payload));
